+++ src/app/shared/module/poe/type/content.type.ts
@@ -1,4 +1,5 @@
 import { ItemCategory } from './item.type';
+import { Stat } from './stat.type';
 
 export interface ClientStringMap {
     [id: string]: string;
@@ -8,20 +9,14 @@ export interface WordMap {
     [id: string]: string;
 }
 
-export interface StatsDescriptionMap {
-    [keys: string]: {
-        [predicate: string]: string;
-    };
-}
-
-export interface StatsIdMap {
-    [id: string]: string;
-}
-
 export interface BaseItemTypeMap {
     [id: string]: string;
 }
 
 export interface BaseItemTypeCategoryMap {
     [id: string]: ItemCategory;
+}
+
+export interface StatMap {
+    [id: string]: Stat;
 }