+++ src/app/shared/module/poe/type/index.ts
@@ -4,4 +4,5 @@ export * from './currency.type';
 export * from './item.type';
 export * from './language.type';
 export * from './league.type';
+export * from './stat.type';