+++ src/app/shared/module/poe/type/stat.type.ts
@@ -1,0 +1,19 @@
+export enum StatType {
+    Pseudo = 'pseudo',
+    Explicit = 'explicit',
+    Implicit = 'implicit',
+    Crafted = 'crafted',
+    Fractured = 'fractured',
+    Enchant = 'enchant',
+    Veiled = 'veiled',
+    Monster = 'monster',
+    Delve = 'delve'
+}
+
+export interface Stat {
+    id?: string;
+    mod?: string;
+    text: {
+        [language: number]: string;
+    };
+}